+++ flash_attn/cute/softmax.py
@@ -8,27 +8,6 @@ from dataclasses import dataclass
 import cutlass
 import cutlass.cute as cute
 
-# =============================================================================
-# Forward-dropout HW/SW exp2 schedule constants.
-#
-# Both values are sweep winners on the canonical
-# B=4,S=8K,Hq=8,Hk=1,D=128 config; every other setting regressed, so
-# there is no runtime tuning lever.
-# =============================================================================
-
-# 8-bit bitmask selecting which g-slots in ``apply_exp2_convert_row_sum_pair``
-# use a software polynomial exp2 (FMA pipe via ``utils.ex2_emulation_2``)
-# vs the hardware ``MUFU.EX2`` (XU pipe). ``4`` = g=2 SW fills one XU-pipe
-# scheduling bubble between ``MUFU.EX2`` and the dependent ``FSEL`` / ``F2FP``
-# work, winning ~3 us.
-FA4_DROP_E2E_MASK: int = 4
-
-# Polynomial degree for the software exp2 emulator used at the SW slot(s)
-# above. ``1`` (= 2 terms) is the winner of a 0..5 sweep; higher degrees
-# re-open the scheduling bubble and regress.
-FA4_DROP_E2E_POLY_DEGREE: int = 1
-
-
 from cutlass import Float32, Int32, Boolean
 
 from quack import layout_utils
@@ -38,460 +17,15 @@ from flash_attn.cute.seqlen_info import SeqlenInfoQK
 from flash_attn.cute.philox import (
     philox_unpack_seed_offset,
     philox_rounds,
-    philox_precompute_round_keys,
-    philox_rounds_with_keys,
 )
-from cutlass.cutlass_dsl import dsl_user_op
-from cutlass.cute.typing import Int32 as Int32Type, Int64, Uint32, Uint64
-from cutlass._mlir import ir
-from cutlass._mlir.dialects import arith, llvm
-
-
-@dsl_user_op
-def f32_to_dropout_threshold(p_dropout: Float32, *, loc=None, ip=None) -> Uint32:
-    """Convert float dropout probability to packed 8-bit threshold for f16x2 comparison.
-
-    Returns (uint8(p * 255) << 16) | uint8(p * 255).
-    """
-    val = Float32(p_dropout).ir_value(loc=loc, ip=ip)
-    i32_ty = ir.IntegerType.get_signless(32)
-    f32_ty = ir.F32Type.get()
-
-    scale = arith.constant(f32_ty, 255.0, loc=loc, ip=ip)
-    scaled = arith.mulf(val, scale, loc=loc, ip=ip)
-    uint8_val = arith.fptoui(i32_ty, scaled, loc=loc, ip=ip)
-
-    sixteen = arith.constant(i32_ty, 16, loc=loc, ip=ip)
-    hi = arith.shli(uint8_val, sixteen, loc=loc, ip=ip)
-    packed = arith.ori(hi, uint8_val, loc=loc, ip=ip)
-
-    return Uint32(packed)
-
-
-@dsl_user_op
-def dropout_mask_to_bit(mask_val: Float32, bit_pos: Uint32, accum: Uint32, *, loc=None, ip=None) -> Uint32:
-    """Pack a Float32 dropout mask (0.0 or 1.0) into a bit at bit_pos in accum."""
-    i32_ty = ir.IntegerType.get_signless(32)
-    f32_ty = ir.F32Type.get()
-    val = Float32(mask_val).ir_value(loc=loc, ip=ip)
-    half = arith.constant(f32_ty, 0.5, loc=loc, ip=ip)
-    keep = arith.cmpf(arith.CmpFPredicate.OGT, val, half, loc=loc, ip=ip)
-    one = arith.constant(i32_ty, 1, loc=loc, ip=ip)
-    zero = arith.constant(i32_ty, 0, loc=loc, ip=ip)
-    bit = arith.select(keep, one, zero, loc=loc, ip=ip)
-    pos = Uint32(bit_pos).ir_value(loc=loc, ip=ip)
-    shifted = arith.shli(bit, pos, loc=loc, ip=ip)
-    acc = Uint32(accum).ir_value(loc=loc, ip=ip)
-    result = arith.ori(acc, shifted, loc=loc, ip=ip)
-    return Uint32(result)
-
-
-@dsl_user_op
-def dropout_bit_to_scale(bits: Uint32, bit_pos: Uint32, scale: Float32, *, loc=None, ip=None) -> Float32:
-    """Extract bit at bit_pos from bits. Return scale if 1, else 0.0."""
-    i32_ty = ir.IntegerType.get_signless(32)
-    f32_ty = ir.F32Type.get()
-    bits_val = Uint32(bits).ir_value(loc=loc, ip=ip)
-    pos = Uint32(bit_pos).ir_value(loc=loc, ip=ip)
-    shifted = arith.shrui(bits_val, pos, loc=loc, ip=ip)
-    one = arith.constant(i32_ty, 1, loc=loc, ip=ip)
-    bit = arith.andi(shifted, one, loc=loc, ip=ip)
-    zero_i = arith.constant(i32_ty, 0, loc=loc, ip=ip)
-    is_keep = arith.cmpi(arith.CmpIPredicate.ne, bit, zero_i, loc=loc, ip=ip)
-    scale_val = Float32(scale).ir_value(loc=loc, ip=ip)
-    zero_f = arith.constant(f32_ty, 0.0, loc=loc, ip=ip)
-    return Float32(arith.select(is_keep, scale_val, zero_f, loc=loc, ip=ip))
-
-
-@dsl_user_op
-def f32_to_dropout_threshold_byte(p_dropout: Float32, *, loc=None, ip=None) -> Uint32:
-    """Convert float dropout probability to single threshold byte as Uint32."""
-    val = Float32(p_dropout).ir_value(loc=loc, ip=ip)
-    i32_ty = ir.IntegerType.get_signless(32)
-    f32_ty = ir.F32Type.get()
-    scale = arith.constant(f32_ty, 255.0, loc=loc, ip=ip)
-    scaled = arith.mulf(val, scale, loc=loc, ip=ip)
-    return Uint32(arith.fptoui(i32_ty, scaled, loc=loc, ip=ip))
-
-
-@dsl_user_op
-def philox_extract_mask_bits_fa2(
-    rng0: Uint32, rng1: Uint32, rng2: Uint32, rng3: Uint32,
-    threshold_byte: Uint32,
-    byte_pair_offset: Uint32,
-    sub_lane: Uint32,
-    *, loc=None, ip=None
-) -> Uint32:
-    """Extract 8 mask bits from one FA2-convention Philox call and place them
-    at the correct positions within a 32-bit mask word.
-
-    For each byte_group bg (0-3), extracts 2 bytes at byte_pair_offset and
-    byte_pair_offset+1 from rng_words[bg], compares against threshold, and
-    places the 2 result bits at positions bg*8 + sub_lane*2 and bg*8 + sub_lane*2 + 1
-    in the output word.
-    """
-    i32_ty = ir.IntegerType.get_signless(32)
-    r = [Uint32(x).ir_value(loc=loc, ip=ip) for x in [rng0, rng1, rng2, rng3]]
-    th = Uint32(threshold_byte).ir_value(loc=loc, ip=ip)
-    bpo = Uint32(byte_pair_offset).ir_value(loc=loc, ip=ip)
-    sl = Uint32(sub_lane).ir_value(loc=loc, ip=ip)
-
-    asm_lines = []
-    asm_lines.append(
-        "{ .reg .u32 b, sh0, sh1, base, res; .reg .pred p;"
-        " mov.u32 res, 0;"
-        " shl.b32 sh0, $6, 3;"
-        " add.u32 sh1, sh0, 8;"
-        " shl.b32 base, $7, 1;"
-    )
-    regs = ["$1", "$2", "$3", "$4"]
-    for bg in range(4):
-        bit_even = bg * 8
-        bit_odd = bg * 8 + 1
-        asm_lines.append(
-            f" shr.b32 b, {regs[bg]}, sh0; and.b32 b, b, 255;"
-            f" setp.le.u32 p, b, $5; @p or.b32 res, res, {1 << bit_even};"
-            f" shr.b32 b, {regs[bg]}, sh1; and.b32 b, b, 255;"
-            f" setp.le.u32 p, b, $5; @p or.b32 res, res, {1 << bit_odd};"
-        )
-    asm_lines.append(" shl.b32 res, res, base; mov.u32 $0, res; }")
-    asm_str = "".join(asm_lines)
-
-    result = llvm.inline_asm(
-        i32_ty,
-        [*r, th, bpo, sl],
-        asm_str,
-        "=r,r,r,r,r,r,r,r",
-        has_side_effects=False,
-        is_align_stack=False,
-        asm_dialect=llvm.AsmDialect.AD_ATT,
-        loc=loc,
-        ip=ip,
-    )
-    return Uint32(result)
-
-
-@dsl_user_op
-def philox_extract_mask_16bits(
-    rng0: Uint32, rng1: Uint32, rng2: Uint32, rng3: Uint32,
-    threshold_byte: Uint32,
-    bit_offset: Uint32,
-    *, loc=None, ip=None
-) -> Uint32:
-    """Extract 16 mask bits from one Philox call using ALL 4 bytes per word.
-
-    Uses the optimized convention: each rng word's 4 bytes produce 4 consecutive
-    mask bits. rng0 -> bits 0-3, rng1 -> bits 4-7, rng2 -> bits 8-11, rng3 -> bits 12-15.
-    The 16 result bits are shifted left by bit_offset before returning.
-    """
-    i32_ty = ir.IntegerType.get_signless(32)
-    r = [Uint32(x).ir_value(loc=loc, ip=ip) for x in [rng0, rng1, rng2, rng3]]
-    th = Uint32(threshold_byte).ir_value(loc=loc, ip=ip)
-    boff = Uint32(bit_offset).ir_value(loc=loc, ip=ip)
-
-    asm_lines = []
-    asm_lines.append(
-        "{ .reg .u32 b, res; .reg .pred p;"
-        " mov.u32 res, 0;"
-    )
-    regs = ["$1", "$2", "$3", "$4"]
-    for w in range(4):
-        for byte_idx in range(4):
-            bit_pos = w * 4 + byte_idx
-            shift = byte_idx * 8
-            asm_lines.append(
-                f" shr.b32 b, {regs[w]}, {shift}; and.b32 b, b, 255;"
-                f" setp.le.u32 p, b, $5; @p or.b32 res, res, {1 << bit_pos};"
-            )
-    asm_lines.append(" shl.b32 res, res, $6; mov.u32 $0, res; }")
-    asm_str = "".join(asm_lines)
-
-    result = llvm.inline_asm(
-        i32_ty,
-        [*r, th, boff],
-        asm_str,
-        "=r,r,r,r,r,r,r",
-        has_side_effects=False,
-        is_align_stack=False,
-        asm_dialect=llvm.AsmDialect.AD_ATT,
-        loc=loc,
-        ip=ip,
-    )
-    return Uint32(result)
-
-
-@dsl_user_op
-def smem_mask_store_u32(
-    smem_ptr: cute.Pointer, offset: Int32Type, mask_word: Uint32,
-    *, loc=None, ip=None
-) -> None:
-    """Store a 32-bit mask word to shared memory at smem_ptr + offset * 4."""
-    ptr_val = smem_ptr.toint(loc=loc, ip=ip).ir_value()
-    off_val = Int32Type(offset).ir_value(loc=loc, ip=ip)
-    w_val = Uint32(mask_word).ir_value(loc=loc, ip=ip)
-    i32_ty = ir.IntegerType.get_signless(32)
-
-    llvm.inline_asm(
-        i32_ty,
-        [ptr_val, off_val, w_val],
-        "{ .reg .u64 off64, addr;"
-        " cvt.u64.s32 off64, $2;"
-        " shl.b64 off64, off64, 2;"
-        " add.u64 addr, $1, off64;"
-        " st.shared.u32 [addr], $3; }",
-        "=r,l,r,r",
-        has_side_effects=True,
-        is_align_stack=False,
-        asm_dialect=llvm.AsmDialect.AD_ATT,
-        loc=loc,
-        ip=ip,
-    )
-
-
-@dsl_user_op
-def smem_mask_store_combined(
-    smem_ptr: cute.Pointer, offset: Int32Type, bits_lo: Uint32, bits_hi: Uint32,
-    *, loc=None, ip=None
-) -> None:
-    """Store (bits_hi << 16) | bits_lo to shared memory at smem_ptr + offset * 4."""
-    ptr_val = smem_ptr.toint(loc=loc, ip=ip).ir_value()
-    off_val = Int32Type(offset).ir_value(loc=loc, ip=ip)
-    lo_val = Uint32(bits_lo).ir_value(loc=loc, ip=ip)
-    hi_val = Uint32(bits_hi).ir_value(loc=loc, ip=ip)
-    i32_ty = ir.IntegerType.get_signless(32)
-
-    llvm.inline_asm(
-        i32_ty,
-        [ptr_val, off_val, lo_val, hi_val],
-        "{ .reg .u64 off64, addr; .reg .u32 combined;"
-        " shl.b32 combined, $4, 16;"
-        " or.b32 combined, combined, $3;"
-        " cvt.u64.s32 off64, $2;"
-        " shl.b64 off64, off64, 2;"
-        " add.u64 addr, $1, off64;"
-        " st.shared.u32 [addr], combined; }",
-        "=r,l,r,r,r",
-        has_side_effects=True,
-        is_align_stack=False,
-        asm_dialect=llvm.AsmDialect.AD_ATT,
-        loc=loc,
-        ip=ip,
-    )
-
-
-@dsl_user_op
-def smem_mask_load_u32(
-    smem_ptr: cute.Pointer, offset: Int32Type,
-    *, loc=None, ip=None
-) -> Uint32:
-    """Load a uint32 from shared memory at smem_ptr + offset * 4."""
-    ptr_val = smem_ptr.toint(loc=loc, ip=ip).ir_value()
-    off_val = Int32Type(offset).ir_value(loc=loc, ip=ip)
-    i32_ty = ir.IntegerType.get_signless(32)
-
-    # has_side_effects=True is REQUIRED here. Without it, the compiler treats
-    # this load as a pure function and may CSE/hoist it across barriers,
-    # causing the consumer to read stale producer data (the very first
-    # producer iteration's bits) on every iteration.
-    result = llvm.inline_asm(
-        i32_ty,
-        [ptr_val, off_val],
-        "{ .reg .u64 off64, addr;"
-        " cvt.u64.s32 off64, $2;"
-        " shl.b64 off64, off64, 2;"
-        " add.u64 addr, $1, off64;"
-        " ld.shared.u32 $0, [addr]; }",
-        "=r,l,r",
-        has_side_effects=True,
-        is_align_stack=False,
-        asm_dialect=llvm.AsmDialect.AD_ATT,
-        loc=loc,
-        ip=ip,
-    )
-    return Uint32(result)
-
-
-@dsl_user_op
-def store_dropout_mask_u8(
-    mask_ptr: cute.Pointer, col_offset: Int32Type, mask_val: Float32,
-    *, loc=None, ip=None
-) -> None:
-    """Store a Float32 dropout mask (0.0 or 1.0) as uint8 to global memory."""
-    from cutlass.cute.typing import Int32 as _Int32
-    ptr_val = mask_ptr.toint(loc=loc, ip=ip).ir_value()
-    off_val = _Int32(col_offset).ir_value(loc=loc, ip=ip)
-    mask_ir = Float32(mask_val).ir_value(loc=loc, ip=ip)
-
-    i32_ty = ir.IntegerType.get_signless(32)
-
-    llvm.inline_asm(
-        i32_ty,
-        [ptr_val, off_val, mask_ir],
-        "{ .reg .u64 off64, addr; .reg .u32 v;"
-        " cvt.u64.s32 off64, $2;"
-        " add.u64 addr, $1, off64;"
-        " cvt.rzi.u32.f32 v, $3;"
-        " st.global.u8 [addr], v; }",
-        "=r,l,r,f",
-        has_side_effects=True,
-        is_align_stack=False,
-        asm_dialect=llvm.AsmDialect.AD_ATT,
-        loc=loc,
-        ip=ip,
-    )
-
-
-@dsl_user_op
-def prmt_b32(rng: Uint32, b: Uint32, sel: Uint32, *, loc=None, ip=None) -> Uint32:
-    """PTX ``prmt.b32 dst, rng, b, sel``: byte-permute (rng||b) per ``sel``.
-
-    ``sel`` is a 32-bit value with 4 nibbles (low to high = byte 0..3 of dst).
-    Each nibble's lower 3 bits select one of the 8 source bytes from
-    ``(rng[0..3], b[0..3])``; bit 3 enables sign-extension from the MSB of
-    the selected byte. To extract byte ``i`` of ``rng`` into the LSB of the
-    result with the other bytes zeroed, pass ``b=0`` and ``sel=0x4440 | i``.
-    """
-    i32_ty = ir.IntegerType.get_signless(32)
-    rng_ir = Uint32(rng).ir_value(loc=loc, ip=ip)
-    b_ir = Uint32(b).ir_value(loc=loc, ip=ip)
-    sel_ir = Uint32(sel).ir_value(loc=loc, ip=ip)
-    result = llvm.inline_asm(
-        i32_ty,
-        [rng_ir, b_ir, sel_ir],
-        "prmt.b32 $0, $1, $2, $3;",
-        "=r,r,r,r",
-        has_side_effects=False,
-        is_align_stack=False,
-        asm_dialect=llvm.AsmDialect.AD_ATT,
-        loc=loc,
-        ip=ip,
-    )
-    return Uint32(result)
-
-
-@dsl_user_op
-def apply_dropout_pair(
-    val_lo: Float32, val_hi: Float32,
-    rng_word: Uint32,
-    sel_lo: Uint32, sel_hi: Uint32,
-    thresh_byte: Uint32,
-    rp_dropout: Float32,
-    *, loc=None, ip=None
-):
-    """Apply dropout + rp_dropout scaling to a pair of Float32 values.
-
-    Returns ``(select(keep_lo, val_lo * rp_dropout, 0),
-               select(keep_hi, val_hi * rp_dropout, 0))`` where
-    ``keep_x = byte_x <= thresh_byte`` and
-    ``byte_x = prmt.b32(rng_word, 0, sel_x) & 0xff``.
-
-    Bit-identical to FA2: ``prmt.b32(rng, 0, 0x4440 | i)`` produces exactly
-    the same byte value as ``(rng >> (i*8)) & 0xff``, and the
-    ``<= threshold`` comparison is unchanged.
-    """
-    val_lo_ir = Float32(val_lo).ir_value(loc=loc, ip=ip)
-    val_hi_ir = Float32(val_hi).ir_value(loc=loc, ip=ip)
-    thresh_byte_ir = Uint32(thresh_byte).ir_value(loc=loc, ip=ip)
-    rp_ir = Float32(rp_dropout).ir_value(loc=loc, ip=ip)
-
-    f32_ty = ir.F32Type.get()
-    zero_f32 = arith.constant(f32_ty, 0.0, loc=loc, ip=ip)
-    zero_u32 = Uint32(0)
-
-    rnd_lo_u = prmt_b32(rng_word, zero_u32, sel_lo, loc=loc, ip=ip)
-    rnd_hi_u = prmt_b32(rng_word, zero_u32, sel_hi, loc=loc, ip=ip)
-    rnd_lo = Uint32(rnd_lo_u).ir_value(loc=loc, ip=ip)
-    rnd_hi = Uint32(rnd_hi_u).ir_value(loc=loc, ip=ip)
-
-    keep_lo = arith.cmpi(arith.CmpIPredicate.ule, rnd_lo, thresh_byte_ir, loc=loc, ip=ip)
-    keep_hi = arith.cmpi(arith.CmpIPredicate.ule, rnd_hi, thresh_byte_ir, loc=loc, ip=ip)
-
-    scaled_lo = arith.mulf(val_lo_ir, rp_ir, loc=loc, ip=ip)
-    scaled_hi = arith.mulf(val_hi_ir, rp_ir, loc=loc, ip=ip)
-    result_lo = arith.select(keep_lo, scaled_lo, zero_f32, loc=loc, ip=ip)
-    result_hi = arith.select(keep_hi, scaled_hi, zero_f32, loc=loc, ip=ip)
-
-    return (Float32(result_lo), Float32(result_hi))
-
-
-@dsl_user_op
-def apply_dropout_quad_select_only(
-    v0: Float32, v1: Float32, v2: Float32, v3: Float32,
-    rng_word: Uint32,
-    sel_0: Uint32, sel_1: Uint32, sel_2: Uint32, sel_3: Uint32,
-    thresh_byte: Uint32,
-    *, loc=None, ip=None
-):
-    """Dropout select on 4 FP32 values that share one Philox word.
-
-    Computes ``select(byte_i <= thresh_byte, v_i, 0)`` for i in 0..3,
-    where ``byte_i = prmt(rng_word, 0, sel_i)``. Used on the
-    ``.16x256b`` path where one rng_word covers a (top, bot) row-pair
-    times 2 cols == 4 cells, and the per-element ``val * rp`` factor
-    has already been baked into ``acc_S`` by exp2 (see ``log2_rp`` in
-    ``scale_subtract_rowmax_pair`` and ``inv_rp`` in
-    ``apply_exp2_convert_row_sum_pair``), so the select is pure SELP
-    with no FMUL.
-
-    Bit-identical to FA2: byte extraction matches PRMT exactly and
-    the ``<= threshold`` comparison is unchanged.
-    """
-    val_irs = [Float32(v).ir_value(loc=loc, ip=ip) for v in (v0, v1, v2, v3)]
-    thresh_byte_ir = Uint32(thresh_byte).ir_value(loc=loc, ip=ip)
-    sel_irs = [Uint32(s) for s in (sel_0, sel_1, sel_2, sel_3)]
-
-    f32_ty = ir.F32Type.get()
-    zero_f32 = arith.constant(f32_ty, 0.0, loc=loc, ip=ip)
-    zero_u32 = Uint32(0)
-
-    results = []
-    for i in range(4):
-        rnd_byte_u = prmt_b32(rng_word, zero_u32, sel_irs[i], loc=loc, ip=ip)
-        rnd_byte = Uint32(rnd_byte_u).ir_value(loc=loc, ip=ip)
-        keep = arith.cmpi(
-            arith.CmpIPredicate.ule, rnd_byte, thresh_byte_ir, loc=loc, ip=ip
-        )
-        results.append(
-            Float32(arith.select(keep, val_irs[i], zero_f32, loc=loc, ip=ip))
-        )
-
-    return (results[0], results[1], results[2], results[3])
-
-
-@dsl_user_op
-def gen_dropout_mask_f32(
-    rng_word: Uint32, byte_shift: Uint32, threshold_packed: Uint32,
-    *, loc=None, ip=None
-):
-    """Generate a pair of Float32 dropout masks using integer byte comparison."""
-    rng_val = Uint32(rng_word).ir_value(loc=loc, ip=ip)
-    shift_val = Uint32(byte_shift).ir_value(loc=loc, ip=ip)
-    thresh_val = Uint32(threshold_packed).ir_value(loc=loc, ip=ip)
-
-    i32_ty = ir.IntegerType.get_signless(32)
-    f32_ty = ir.F32Type.get()
-
-    mask_0xff = arith.constant(i32_ty, 0xFF, loc=loc, ip=ip)
-    eight = arith.constant(i32_ty, 8, loc=loc, ip=ip)
-    one_f32 = arith.constant(f32_ty, 1.0, loc=loc, ip=ip)
-    zero_f32 = arith.constant(f32_ty, 0.0, loc=loc, ip=ip)
-
-    thresh_byte = arith.andi(thresh_val, mask_0xff, loc=loc, ip=ip)
-
-    shifted_lo = arith.shrui(rng_val, shift_val, loc=loc, ip=ip)
-    rnd_lo = arith.andi(shifted_lo, mask_0xff, loc=loc, ip=ip)
-
-    shift_hi = arith.addi(shift_val, eight, loc=loc, ip=ip)
-    shifted_hi = arith.shrui(rng_val, shift_hi, loc=loc, ip=ip)
-    rnd_hi = arith.andi(shifted_hi, mask_0xff, loc=loc, ip=ip)
-
-    keep_lo = arith.cmpi(arith.CmpIPredicate.ule, rnd_lo, thresh_byte, loc=loc, ip=ip)
-    keep_hi = arith.cmpi(arith.CmpIPredicate.ule, rnd_hi, thresh_byte, loc=loc, ip=ip)
-
-    mask_lo_f32 = arith.select(keep_lo, one_f32, zero_f32, loc=loc, ip=ip)
-    mask_hi_f32 = arith.select(keep_hi, one_f32, zero_f32, loc=loc, ip=ip)
-
-    return (Float32(mask_lo_f32), Float32(mask_hi_f32))
+from flash_attn.cute.dropout import (
+    apply_dropout_pair,
+    apply_dropout_quad_select_only,
+    gen_dropout_mask_f32,
+    store_dropout_mask_u8,
+    fa2_philox_row_index,
+)
+from cutlass.cute.typing import Int64, Uint32, Uint64
 
 
 from flash_attn.cute.utils import AuxData
@@ -977,9 +511,9 @@ class SoftmaxSm100(Softmax):
 
         key_x, key_y, base_ctr_x, base_ctr_y = philox_unpack_seed_offset(rng_seed, rng_offset_adjusted)
 
-        fa2_row_arg = Uint32(actual_seq_row // 16)
-        fa2_base_lane = Uint32((actual_seq_row % 8) * 4)
-        fa2_byte_pair_offset = Uint32(((actual_seq_row // 8) % 2) * 2)
+        fa2_row_arg, fa2_base_lane, fa2_byte_pair_offset = fa2_philox_row_index(
+            actual_seq_row
+        )
 
         n_blk_size = total_cols
         base_col = n_block * n_blk_size
@@ -1321,6 +855,10 @@ class SoftmaxSm100(Softmax):
         group_idx = lane_idx_i32 >> 2  # 0..7
         warp_idx_in_wg = Int32(cute.arch.warp_idx() % (self.m_block_size // 32))
 
+        # Same FA2 key decomposition as ``fa2_philox_row_index`` but hoisted
+        # for the row-pair layout: here ``group_idx == row % 8`` (the other
+        # row terms are multiples of 8), so ``base_lane == (row % 8) * 4`` and
+        # ``fa2_row_arg == top_row // 16`` (computed per row-pair below).
         fa2_base_lane = Uint32(group_idx) * Uint32(4)
 
         sel_top_lo = Uint32(0x4440)
